Add device attribute in Disk class
This attribute hosts the block device associated with this Disk
instance. This can be the original media if the source media is a block
device, a loop device of the original media if the original media is a
regular file or a loop device of a temporary object if the original
media is a directory.