Fix a bug in the VirtIO installation code
Prior to this, if you tried to install the same version of the VirtIO
drivers that where already installed on a Windows media, you would end
up just uninstalling the old ones. This happened because
snf-image-creator added code for installing the new drivers and then for
removing the old ones. Windows would ignore the first request, since the
drivers were already present in the database, but would perform the
second one.

This fixes #44