Keep separate registration status per cloud
Keep track of the cloud accounts an image was uploaded and registered
on. When changing the current cloud account, the registration info
should also change to represent the current cloud's status. We may have
registered an image on an account and then switch to a different
account. The image should not be shown as registered, unless we switch
back to the old one.