windows: Shrink the volume without using querymax
Shrink volume as much as possible and then extend it 100MB. We used to
do a querymax to see how much we could shrink the volume and then
compute the desired shrinking size, but this leaded to problems. We had
to parse the output of "shrink querymax" which is in a localized natural
language format.

This fixes #53