Check for broken links in systemd service files
Make sure the link is not broken before trying to open it, when
inspecting systemd service files. In Ubuntu `snapd.book-ok' and
`snapd.firstboot' service files may be removed and the links found
under `/etc/systemd/system/multi-user.target.wants' may be broken. Prior
to this commit this triggered a RuntimeError.